+++ Happypets - Server/src/main/java/happypets/repository/UserDetailsRepository.java
@@ -3,5 +3,5 @@ package happypets.repository;
 import happypets.model.UserDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserDetailsRepository extends JpaRepository<UserDetails, Long> {
+public interface UserDetailsRepository extends JpaRepository<UserDetails, Integer> {
 }
